Move trig tansform functions to scn.structure.
