+++ sidechainnet/examples/losses.py
@@ -4,7 +4,7 @@ import numpy as np
 import prody as pr
 import torch
 
-from sidechainnet.structure.build_info import NUM_ANGLES, NUM_COORDS_PER_RES
+from sidechainnet.structure.build_info import NUM_COORDS_PER_RES
 
 
 def compute_batch_drmsd(true_coordinates, pred_coordinates, seq, verbose=False):
@@ -49,7 +49,7 @@ def drmsd(a, b):
     Given 2 coordinate tensors, returns the dRMSD between them. Both
     tensors must be the exact same shape. It works by creating a mask of the
     upper-triangular indices of the pairwise distance matrix (excluding the
-    diagonal). Then, the resulting values are compared with Pytorch's MSE loss.
+    diagonal). Then, the resulting values are compared with PyTorch's MSE loss.
 
     Args:
         a, b (torch.Tensor): coordinate tensor with shape (L x 3).
@@ -97,49 +97,6 @@ def pairwise_internal_dist(x):
                       alpha=-2).add_(x1_norm)
     res = res.clamp_min_(1e-30).sqrt_()
     return res
-
-
-def inverse_trig_transform(t):
-    """Compute the atan2 of the last 2 dimensions of a given tensor.
-
-    Given a (BATCH x L X NUM_PREDICTED_ANGLES ) tensor, returns (BATCH X
-    L X NUM_PREDICTED_ANGLES) tensor. Performs atan2 transformation from sin
-    and cos values.
-
-    Args:
-        t (torch.tensor): Tensor of shape (batch, L, NUM_ANGLES, 2).
-
-    Returns:
-        torch.tensor: Tensor of angles with the last two dimensions reduced
-        via atan2.
-    """
-    t = t.view(t.shape[0], -1, NUM_ANGLES, 2)
-    t_cos = t[:, :, :, 0]
-    t_sin = t[:, :, :, 1]
-    t = torch.atan2(t_sin, t_cos)
-    return t
-
-
-def trig_transform(t):
-    """Expand the last dimension of an angle tensor to have sin/cos values.
-
-    Args:
-        t (torch.tensor): Angle tensor with shape (batch x L x num_angle)
-
-    Raises:
-        ValueError: if tensor shape is not recognized.
-
-    Returns:
-        torch.tensor: Angle tensor with shape (batch x L x num_angle x 2).
-    """
-    new_t = torch.zeros(*t.shape, 2)
-    if len(new_t.shape) == 4:
-        new_t[:, :, :, 0] = torch.cos(t)
-        new_t[:, :, :, 1] = torch.sin(t)
-    else:
-        raise ValueError("trig_transform function is only defined for "
-                         "(batch x L x num_angle) tensors.")
-    return new_t
 
 
 def _tile(a, dim, n_tile):
