+++ sidechainnet/structure/__init__.py
@@ -1,0 +1,2 @@
+from .StructureBuilder import StructureBuilder
+from .structure import inverse_trig_transform, trig_transform
