+++ sidechainnet/structure/structure.py
@@ -3,13 +3,12 @@
 import numpy as np
 import torch
 from sidechainnet.structure import StructureBuilder
+from sidechainnet.structure.build_info import NUM_ANGLES
 from sidechainnet.utils.sequence import VOCAB
 
 
 def angles_to_coords(angles, seq, remove_batch_padding=False):
-    """
-    Convert torsional angles to coordinates.
-    """
+    """Convert torsional angles to coordinates."""
     pred_ang, input_seq = angles, seq
     if remove_batch_padding:
         # Remove batch-level masking
@@ -93,6 +92,49 @@ def deg2rad(angle):
     Converts an angle in degrees to radians.
     """
     return angle * np.pi / 180.
+
+
+def inverse_trig_transform(t):
+    """Compute the atan2 of the last 2 dimensions of a given tensor.
+
+    Given a (BATCH x L X NUM_PREDICTED_ANGLES ) tensor, returns (BATCH X
+    L X NUM_PREDICTED_ANGLES) tensor. Performs atan2 transformation from sin
+    and cos values.
+
+    Args:
+        t (torch.tensor): Tensor of shape (batch, L, NUM_ANGLES, 2).
+
+    Returns:
+        torch.tensor: Tensor of angles with the last two dimensions reduced
+        via atan2.
+    """
+    t = t.view(t.shape[0], -1, NUM_ANGLES, 2)
+    t_cos = t[:, :, :, 0]
+    t_sin = t[:, :, :, 1]
+    t = torch.atan2(t_sin, t_cos)
+    return t
+
+
+def trig_transform(t):
+    """Expand the last dimension of an angle tensor to have sin/cos values.
+
+    Args:
+        t (torch.tensor): Angle tensor with shape (batch x L x num_angle)
+
+    Raises:
+        ValueError: if tensor shape is not recognized.
+
+    Returns:
+        torch.tensor: Angle tensor with shape (batch x L x num_angle x 2).
+    """
+    new_t = torch.zeros(*t.shape, 2)
+    if len(new_t.shape) == 4:
+        new_t[:, :, :, 0] = torch.cos(t)
+        new_t[:, :, :, 1] = torch.sin(t)
+    else:
+        raise ValueError("trig_transform function is only defined for "
+                         "(batch x L x num_angle) tensors.")
+    return new_t
 
 
 def debug_example():
